sid-app/pyproject.toml
[tool.poetry]
name = "sid-app"
version = "0.1.0"
description = ""
authors = ["limdongjin <geniuslim27@gmail.com>"]
readme = "README.md"
packages = [{include = "sid"}]

[tool.poetry.dependencies]
python = "^3.9"
# pyignite = "^0.6.0"
confluent-kafka = "^2.0.2"
protobuf = "3.20.3"
librosa = "0.9.2"
# https://download.pytorch.org/whl/cpu
torch = [
    { url = "https://download.pytorch.org/whl/cpu/torch-1.12.0%2Bcpu-cp39-cp39-linux_x86_64.whl", markers = "sys_platform == 'linux'" }
]
torchaudio = [
    { url = "https://download.pytorch.org/whl/cpu/torchaudio-0.12.0%2Bcpu-cp39-cp39-linux_x86_64.whl", markers="sys_platform == 'linux'" }
]
onnx = "1.13.0"
onnxruntime = "1.14.0"
tqdm = "4.64.1"
scipy = "1.9.3"
google-cloud-datastore = "^2.13.2"
pydub = "^0.25.1"
openai = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
